gateway/requirements.txt
# G.O.S. Gateway Services - Python Dependencies
# For Docker container

# Database
psycopg2-binary>=2.9

# Data Processing
pandas>=2.0
numpy>=1.24

# Validation
pydantic>=2.0

# MQTT (for real hardware)
paho-mqtt>=2.0

# API
flask>=3.0
flask-cors>=4.0

# Utilities
python-dotenv>=1.0
requests>=2.31
